requirements.txt
tensorflow
matplotlib
Pillow
numpy
